pyproject.toml
[project]
name = "photokit"
version = "0.1.0"
description = "Command-line image filtering tool — Git & GitHub workshop starter"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Workshop Team" },
]
dependencies = [
    "pillow>=10.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.scripts]
photokit = "photokit.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/photokit"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
